Gabriela Ciurea - Homework 2
Homework 2 completed
