+++ homework/TaskList.js
@@ -1,37 +1,5 @@
-var Task = /** @class */ (function () {
-    function Task(name, endDate, description) {
-        this._isFinished = false;
-        this.name = name;
-        this.endDate = endDate;
-        this.description = description;
-    }
-    Task.prototype.finishTask = function () {
-        this._isFinished = true;
-    };
-    Task.prototype.modifyTask = function (propertyName, propertyValue) {
-        if (this._isFinished) {
-            console.log('The task is finished and it can not be modified');
-            return;
-        }
-        var taskKeys = Object.keys(this);
-        for (var _i = 0, taskKeys_1 = taskKeys; _i < taskKeys_1.length; _i++) {
-            var key = taskKeys_1[_i];
-            if (key === propertyName) {
-                this[propertyName] = propertyValue;
-                return;
-            }
-        }
-        console.log("Property " + propertyName + " does not exist");
-    };
-    Object.defineProperty(Task.prototype, "isFinished", {
-        get: function () {
-            return this._isFinished;
-        },
-        enumerable: true,
-        configurable: true
-    });
-    return Task;
-}());
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
 var TaskList = /** @class */ (function () {
     function TaskList() {
         this.taskList = [];
@@ -70,17 +38,4 @@ var TaskList = /** @class */ (function () {
     };
     return TaskList;
 }());
-var taskList = new TaskList();
-var task1 = new Task('Christmas party', new Date('2019-12-25'), 'description 1');
-var task2 = new Task('First Exam', new Date('2020-01-19'), 'description 2');
-var task3 = new Task('Mothers Day', new Date('2020-03-08'), 'description 3');
-taskList.addTask(task1);
-taskList.addTask(task2);
-taskList.addTask(task3);
-taskList.finishTask(task2);
-taskList.printAll();
-taskList.modifyTask(task1, 'name', 'Task Updated');
-taskList.printAll();
-taskList.deleteTask(task3);
-taskList.printAll();
-taskList.printAllUnfinished();
+exports.TaskList = TaskList;
